# src/mdm.api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "行動裝置管理系統 API — 裝置管控、資產管理、借用流程、App 派發",
        "title": "MDM Server API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/profiles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "描述檔列表 / 上傳描述檔",
                "parameters": [
                    {
                        "type": "file",
                        "description": "mobileconfig 檔案（POST）",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "描述檔名稱（POST）",
                        "name": "name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {profiles: [...]}, POST: {id, name, filename, size}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/controller.swagProfileItem"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "描述檔列表 / 上傳描述檔",
                "parameters": [
                    {
                        "type": "file",
                        "description": "mobileconfig 檔案（POST）",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "描述檔名稱（POST）",
                        "name": "name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {profiles: [...]}, POST: {id, name, filename, size}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/controller.swagProfileItem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "取得描述檔內容（base64）/ 刪除描述檔",
                "parameters": [
                    {
                        "type": "string",
                        "description": "描述檔 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET 回傳 base64 內容",
                        "schema": {
                            "$ref": "#/definitions/controller.swagProfileContentResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "取得描述檔內容（base64）/ 刪除描述檔",
                "parameters": [
                    {
                        "type": "string",
                        "description": "描述檔 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET 回傳 base64 內容",
                        "schema": {
                            "$ref": "#/definitions/controller.swagProfileContentResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.swagError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "controller.swagProfileContentResp": {
            "type": "object",
            "properties": {
                "content_base64": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                }
            }
        },
        "controller.swagProfileItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "uploaded_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}